Pat and Norman's work on backend
Among some other changes, Norman and I renamed the "PongService" to
"GameService" so that this leaderboard is more generalizable.  We also
added a "Player" class as a separate file, but we will still need to
move the functionality of the old "Player" class inside of "GameService" to the new, separate
one.  We also began to write some tests in "PlayerTest" and "MainTest,"
but they will need more tests as well.