Pat and Norman's work on backend
Among some other changes, Norman and I renamed the "PongService" to
"GameService" so that this leaderboard is more generalizable.  We had
thought about moving the "Player" class from inside the "GameService" file to
its own file so that it wouldn't be so tightly coupled with the former.
This might help in making the leaderboard generalizable to other sports
or activities beyond ping pong.  But we're still trying to figure out the best way to store the name of the activity/sport.
Some players will participate in multiple sports/activities, while others
will likely only participate in one.  So we'll still need to add
functionality for a user of the leaderboard to be able to only show the
leaderboard for a particular sport/activity (such as volleyball). They
should then only see the rankings of players who play volleyball, not
necessarily all the players in the system.